Moved all constants in config folder into one file, named index.js, to minimize available files and order the folder structure
