+++ src/api/searchMovies.js
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import { BASE_URL } from '../common/constants/config/BASE_URL';
+import { BASE_URL } from '../common/constants/config';
 // import { API_KEY } from '../config/API_KEY';
 
 const API_KEY = `eyJhbGciOiJIUzI1NiJ9.eyJhdWQiOiIxZDQ2MTI2MDcwMmJkZGJiOTg4MmUyZTRhMDJlZDA0ZSIsIm5iZiI6MTcyMzY2MTY0My45NDk0MjksInN1YiI6IjY2YWI4MjEwNGZlNDIxMzEwY2QyY2FlNSIsInNjb3BlcyI6WyJhcGlfcmVhZCJdLCJ2ZXJzaW9uIjoxfQ.9nMPyeWUKdhaUTVPfltvi473upjSJlz1iKauLIeuXpQ`;
