+++ src/common/components/Tile/index.js
@@ -12,11 +12,11 @@ import {
 } from "./styled";
 import { StyledProfileIcon } from "../StyledProfileIcon";
 import { StyledVideoIcon } from "../StyledVideoIcon";
-import { BASE_IMAGE_URL } from "../../constants/config/BASE_IMAGE_URL";
+import { BASE_IMAGE_URL } from "../../constants/config";
 import { getImageUrl } from "../../../common/functions/getImageUrl";
 import { checkIsValidImageUrl } from "../../../common/functions/checkIsValidImageUrl";
-import { INVALID_IMAGE_URL } from "../../constants/config/INVALID_IMAGE_URL";
-import { IMAGE_WIDTH } from "../../constants/config/IMAGE_WIDTH";
+import { INVALID_IMAGE_URL } from "../../constants/config";
+import { IMAGE_WIDTH } from "../../constants/config";
 
 export const Tile = ({ image, title, subInfo, extraContent, id, navigateTo, }) => {
