--- src/common/constants/config/BACKDROP__WIDTH.js
@@ -1,1 +1,0 @@
-export const BACKDROP_WIDTH = 'original';
