--- src/common/constants/config/BASE_IMAGE_URL.js
@@ -1,1 +1,0 @@
-export const BASE_IMAGE_URL = `https://image.tmdb.org/t/p/`;
