--- src/common/constants/config/IMAGE_WIDTH.js
@@ -1,1 +1,0 @@
-export const IMAGE_WIDTH = 'w500';
