--- src/common/constants/config/INVALID_IMAGE_URL.js
@@ -1,4 +1,0 @@
-import { BASE_IMAGE_URL } from "./BASE_IMAGE_URL";
-import { IMAGE_WIDTH } from "./IMAGE_WIDTH";
-
-export const INVALID_IMAGE_URL = `${BASE_IMAGE_URL}${IMAGE_WIDTH}null`;
