--- src/common/constants/config/PHOTO_WIDTH.js
@@ -1,1 +1,0 @@
-export const PHOTO_WIDTH = 'w185';
