+++ src/common/constants/config/index.js
@@ -1,0 +1,9 @@
+export const API_KEY = '57afc837c270fb1a287ab5e06dc4e352';
+export const BASE_URL = 'https://api.themoviedb.org/3'; 
+
+export const BACKDROP_WIDTH = 'original';
+export const IMAGE_WIDTH = 'w500';
+export const PHOTO_WIDTH = 'w185';
+
+export const BASE_IMAGE_URL = `https://image.tmdb.org/t/p/`;
+export const INVALID_IMAGE_URL = `${BASE_IMAGE_URL}${IMAGE_WIDTH}null`;
