+++ src/common/functions/getImageUrl.js
@@ -1,3 +1,3 @@
-import { BASE_IMAGE_URL } from "../constants/config/BASE_IMAGE_URL";
+import { BASE_IMAGE_URL } from "../constants/config";
 
 export const getImageUrl = (imagePath, width) => `${BASE_IMAGE_URL}${width}${imagePath}`;
