+++ src/features/ActorPage/Content/Details/index.js
@@ -1,5 +1,5 @@
-import { BASE_IMAGE_URL } from "../../../../common/constants/config/BASE_IMAGE_URL";
-import { IMAGE_WIDTH } from "../../../../common/constants/config/IMAGE_WIDTH";
+import { BASE_IMAGE_URL } from "../../../../common/constants/config";
+import { IMAGE_WIDTH } from "../../../../common/constants/config";
 import {
     ActorProfile,
     ActorPhoto,
